# docs/admin-openapi.yaml
openapi: 3.0.3
info:
  title: 只因梗懂你 · 管理端 API
  description: |
    管理端接口（`/admin/**`）及登录、AI 运维相关接口。
    成功响应统一为 `{ code: 1, message, data }`；业务失败抛 BizException，常见 code：0/400/401/403/404。
    鉴权：`Authorization: Bearer <accessToken>`；`/admin/**` 要求 `ROLE_ADMIN`。
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 本地后端
...
paths:
...
  /admin/memes:
    get:
      tags: [AdminMeme]
      summary: 分页查询梗
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - name: status
          in: query
          schema:
            type: integer
            enum: [1, 2, 3]
        - name: keyword
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminMemePage'
  /admin/memes/{id}:
    get:
      tags: [AdminMeme]
      summary: 梗详情
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminMemeDetail'
  /admin/memes/{id}/approve:
    post:
      tags: [AdminMeme]
      summary: 审核通过
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminMemeAction'
  /admin/memes/{id}/reject:
    post:
      tags: [AdminMeme]
      summary: 审核拒绝（下架）
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminMemeAction'
  /admin/memes/{id}/offline:
    post:
      tags: [AdminMeme]
      summary: 强制下架
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminMemeAction'
  /admin/memes/{id}/online:
    post:
      tags: [AdminMeme]
      summary: 重新提交审核
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminMemeAction'
...
  /admin/home-images:
    get:
      tags: [AdminHomeImage]
      summary: 分页查询轮播
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - name: status
          in: query
          schema:
            type: integer
            enum: [0, 1]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminHomeImagePage'
    post:
      tags: [AdminHomeImage]
      summary: 新增轮播
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminHomeImageSaveRequestDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminHomeImage'
  /admin/home-images/{id}:
    put:
      tags: [AdminHomeImage]
      summary: 更新轮播
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminHomeImageSaveRequestDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminHomeImage'
    delete:
      tags: [AdminHomeImage]
      summary: 删除轮播
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      responses:
        '200':
          description: OK
  /admin/home-images/{id}/status:
    patch:
      tags: [AdminHomeImage]
      summary: 上/下线轮播
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: integer
                  enum: [0, 1]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminHomeImage'
...
  /admin/knowledge:
    get:
      tags: [AdminKnowledge]
      summary: 分页查询知识库
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - name: keyword
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminKnowledgePage'
    post:
      tags: [AdminKnowledge]
      summary: 传入知识（MySQL SSOT + 异步 RAG 灌库）
      description: |
        保存到 MySQL `admin_knowledge`，异步调用 MemeAgent `POST /ingest/knowledge`：
        LangChain RecursiveCharacterTextSplitter 切块 → Embedding → pgvector `knowledge_vector_chunk`。
        Agent 可通过工具 `search_admin_knowledge` 检索。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminKnowledgeSaveRequestDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminKnowledge'
  /admin/knowledge/upload:
    post:
      tags: [AdminKnowledge]
      summary: 上传 Markdown / PDF 并灌库
      description: |
        multipart/form-data：`file`（必填，.md/.markdown/.pdf，≤15MB）、
        `title`（可选，默认文件名）、`category`、`tags`（逗号分隔）。
        服务端提取文本后写入 MySQL 并异步向量灌库。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                category:
                  type: string
                tags:
                  type: string
                  description: 逗号分隔标签
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminKnowledge'
  /admin/knowledge/{id}:
    get:
      tags: [AdminKnowledge]
      summary: 知识详情
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminKnowledge'
    put:
      tags: [AdminKnowledge]
      summary: 更新知识并重新灌库
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminKnowledgeSaveRequestDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminKnowledge'
    delete:
      tags: [AdminKnowledge]
      summary: 删除知识（软删 + 删向量）
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      responses:
        '200':
          description: OK
  /admin/knowledge/{id}/reindex:
    post:
      tags: [AdminKnowledge]
      summary: 强制重新灌库
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultAdminKnowledge'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    size:
      name: size
      in: query
      schema:
        type: integer
        default: 10
        maximum: 50
    pathId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    ResultBase:
      type: object
      properties:
        code:
          type: integer
          example: 1
        message:
          type: string
        data: {}
...
    PageMeta:
      type: object
      properties:
        list:
          type: array
          items: {}
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
        hasMore:
          type: boolean
    AdminMemeListItemVO:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        introduction: { type: string }
        image: { type: string }
        status: { type: integer }
        statusDesc: { type: string }
        userId: { type: string }
        authorNickname: { type: string }
        pageViews: { type: integer }
        likes: { type: integer }
        comments: { type: integer }
        releaseTime: { type: string }
        updateTime: { type: string }
    AdminMemeActionVO:
      type: object
      properties:
        memeId: { type: string }
        status: { type: integer }
        statusDesc: { type: string }
...
    AdminHomeImageVO:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        imgUrl: { type: string }
        targetType: { type: integer }
        targetValue: { type: string }
        sortOrder: { type: integer }
        status: { type: integer }
        createTime: { type: string }
        updateTime: { type: string }
    AdminHomeImageSaveRequestDTO:
      type: object
      required: [imgUrl, targetType]
      properties:
        title: { type: string }
        imgUrl: { type: string }
        targetType: { type: integer, enum: [0, 1, 2] }
        targetValue: { type: string }
        sortOrder: { type: integer }
        status: { type: integer, enum: [0, 1] }
...
    ResultAdminMemePage:
      allOf:
        - $ref: '#/components/schemas/ResultBase'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageMeta'
                - type: object
                  properties:
                    list:
                      type: array
                      items:
                        $ref: '#/components/schemas/AdminMemeListItemVO'
    ResultAdminMemeDetail:
      allOf:
        - $ref: '#/components/schemas/ResultBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminMemeListItemVO'
    ResultAdminMemeAction:
      allOf:
        - $ref: '#/components/schemas/ResultBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminMemeActionVO'
...
    ResultAdminHomeImagePage:
      allOf:
        - $ref: '#/components/schemas/ResultBase'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageMeta'
                - type: object
                  properties:
                    list:
                      type: array
                      items:
                        $ref: '#/components/schemas/AdminHomeImageVO'
    ResultAdminHomeImage:
      allOf:
        - $ref: '#/components/schemas/ResultBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminHomeImageVO'
...
    AdminKnowledgeSaveRequestDTO:
      type: object
      required: [title, content]
      properties:
        title: { type: string, maxLength: 200 }
        content: { type: string, maxLength: 200000 }
        category: { type: string, maxLength: 64 }
        tags:
          type: array
          items: { type: string }
          maxItems: 20
    AdminKnowledgeVO:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        content: { type: string }
        category: { type: string }
        tags:
          type: array
          items: { type: string }
        contentHash: { type: string }
        status: { type: integer }
        createTime: { type: string }
        updateTime: { type: string }
    ResultAdminKnowledgePage:
      allOf:
        - $ref: '#/components/schemas/ResultBase'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageMeta'
                - type: object
                  properties:
                    list:
                      type: array
                      items:
                        $ref: '#/components/schemas/AdminKnowledgeVO'
    ResultAdminKnowledge:
      allOf:
        - $ref: '#/components/schemas/ResultBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminKnowledgeVO'